hidl: fingerprint: Add workaround for FPC1020-based fingerprint HALs
On devices shipping with the FPC1020 fingerprint sensor, LG modifies
the HAL to add a single u32 in the middle of the fingerprint_msg_t
struct. This broke fingerprint completely when using the default
open-source fingerprint HIDL service.

Thus, add our own version of the fingerprint header from libhardware
(as of android-14.0.0_r66), add the variable, and guard it with a Soong
config module.

Change-Id: I96a783729031a64c47676ec51b917d978a9b5ff2