tools: add our custom extract_utils helper script
This helper script allows to extract KDZ files, which are firmware
packages on LGE devices.

It overrides prepare_images() function in the base extract-utils helper
script; it will attempt to use KDZ files first before falling back to
the Android OTA extract function needed in case the source is a
LineageOS zip.

Change-Id: I08fe86ed37339bedf2702a1ae27bc7028b60b21e